requirements.txt
accelerate
datasets
diffusers
matplotlib
numpy
pillow
seaborn
torch
torchvision
transformers
